+++ CHANGELOG.md
@@ -37,11 +37,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- [#1081](https://github.com/alleslabs/celatone-frontend/pull/1081) Add service APIs for move verify module
+
 ### Features
+
+- [#1082](https://github.com/alleslabs/celatone-frontend/pull/1082) Replace CONFIG_CHAINS by fetching the chains from the API
+- [#1077](https://github.com/alleslabs/celatone-frontend/pull/1077) Add download minitia custom JSON
+- [#1074](https://github.com/alleslabs/celatone-frontend/pull/1074) Add custom minitia upload JSON page
+- [#1002](https://github.com/alleslabs/celatone-frontend/pull/1002) Add contract verification modal
 
 ### Improvements
 
+- [#1054](https://github.com/alleslabs/celatone-frontend/pull/1054) Improve main search bar to show both address and address as username
+
 ### Bug fixes
+
+- [#1075](https://github.com/alleslabs/celatone-frontend/pull/1075) Fix custom minitia minor change
 
 ## v1.7.3